Moved all config functions out into separate datman_utils
