+++ dashboard/utils.py
@@ -5,34 +5,7 @@ import json
 import time
 import logging
 
-import datman.config
-
 logger = logging.getLogger(__name__)
-
-
-def get_study_path(study, folder=None):
-    """
-    Returns the full path to the study on the file system.
-
-    If folder is supplied and is defined in study config
-    then path to the folder is returned instead.
-    """
-    cfg = datman.config.config()
-    if folder:
-        try:
-            path = cfg.get_path(folder, study)
-        except Exception as e:
-            logger.error("Failed to find folder {} for study {}. Reason: {}"
-                         "".format(folder, study, e))
-            path = None
-        return path
-
-    try:
-        path = cfg.get_study_base(study=study)
-    except Exception as e:
-        logger.error("Failed to find path for {}. Reason: {}".format(study, e))
-        path = None
-    return path
 
 
 def read_json(json_file):
@@ -56,29 +29,3 @@ def get_software_version(json_contents):
     except KeyError:
         software_version = "Version Not Available"
     return software_name + " - " + software_version
-
-
-def update_header_diffs(scan):
-    site = scan.session.timepoint.site_id
-    config = datman.config.config(study=scan.get_study().id)
-
-    try:
-        tolerance = config.get_key("HeaderFieldTolerance", site=site)
-    except Exception:
-        tolerance = {}
-    try:
-        ignore = config.get_key("IgnoreHeaderFields", site=site)
-    except Exception:
-        ignore = []
-
-    tags = config.get_tags(site=site)
-    try:
-        qc_type = tags.get(scan.tag, "qc_type")
-    except KeyError:
-        check_bvals = False
-    else:
-        check_bvals = qc_type == 'dti'
-
-    scan.update_header_diffs(ignore=ignore,
-                             tolerance=tolerance,
-                             bvals=check_bvals)
